pyproject.toml
[project]
name = "django-experiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django==5.2",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "gunicorn>=23.0.0",
    "markdown>=3.8.2",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.5",
]
